# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Costs Service",
    "description": "Cost tracking and budget management service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3011"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "CostName": {
        "type": "string",
        "description": "Provider cost identifier ({provider}-{service-or-model}-{unit-type})",
        "example": "anthropic-sonnet-4.5-tokens-input"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Price": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "pricePerUnitInUsdCents": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "providerDomain": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "pricingRegime": {
            "type": "string",
            "nullable": true,
            "description": "Time-of-day pricing regime this price belongs to ('peak' or 'off-peak'), or null when the provider charges one rate at every hour. The cost name carries the same segment, so a consumer picks the name for a moment rather than computing a rate.",
            "example": "peak"
          },
          "regimeHoursUtc": {
            "type": "string",
            "nullable": true,
            "description": "UTC windows during which this regime is in force, as comma-separated half-open HH:MM-HH:MM ranges (a window may wrap past 24:00 only as the literal end 24:00). Null when pricingRegime is null. A provider's regimes partition the day, so for a given model and token class exactly one cost name matches any instant.",
            "example": "01:00-04:00,06:00-10:00"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "pricePerUnitInUsdCents",
          "provider",
          "providerDomain",
          "type",
          "unit",
          "pricingRegime",
          "regimeHoursUtc",
          "effectiveFrom"
        ]
      },
...
    },
    "parameters": {
      "CostName": {
        "schema": {
          "$ref": "#/components/schemas/CostName"
        },
        "required": true,
        "description": "Provider cost identifier ({provider}-{service-or-model}-{unit-type})",
        "name": "name",
        "in": "path"
      },
...
    }
  },
  "paths": {
...
    "/v1/platform-prices": {
      "get": {
        "operationId": "listPlatformPrices",
        "summary": "List current platform prices for all cost names (public, no identity headers required)",
        "responses": {
          "200": {
            "description": "Current prices resolved via platform plan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Price"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platform-prices/{name}": {
      "get": {
        "operationId": "getPlatformPrice",
        "summary": "Get current platform price for a cost name (public, no identity headers required)",
        "parameters": [
          {
            "$ref": "#/components/parameters/CostName"
          }
        ],
        "responses": {
          "200": {
            "description": "Current price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Price"
                }
              }
            }
          },
          "404": {
            "description": "Price not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error (e.g. no platform plan configured)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}